Update mcsmrt.rb
Testing a switch from UBLAST algorithm to USEARCH algorithm in MCSMRT main program where OTU centroids are aligned to the database references and the top hit is returned. USEARCH (i.e. usearch local) is likely more suitable for this task and in previous testing has shown more consistent and more accurate results when comparing to the taxonomy classification with UTAX.